Share Validation & Alter features of SolarLunarCalendar
